pyproject.toml
[project]
name = "pic-meme-forwarder"
version = "0.1.0"
description = "Twitter meme forwarder to QQ"
requires-python = ">=3.10"
dependencies = [
    "nonebot2[fastapi]>=2.4.0",
    "nonebot-adapter-onebot>=2.4.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "nonebot-plugin-apscheduler>=0.5.0",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
